VknSpace has a maxDepth, VknVector has sizes rather than last positions or last idx, some debugging lead to the inclusion of a VknVector::exists function, VknSpace has the 'Rule of 5' functions, and the project structure better represents the relationships between all the classes. Saving progress, have more syntax errors to debug.

+++ VknConfig/include/VknEngine.hpp
@@ -144,7 +144,7 @@ namespace vkn
         uint32_t push_back(T val)
         {
             VknVector<T> &vec{this->getVector<T>()};
-            uint32_t pos = vec.getSize();
+            size_t pos = vec.getSize();
             vec.append(val);
             return pos;
         }
